requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
trl>=0.7.4
datasets>=2.14.5
fire>=0.5.0
streamlit>=1.29.0
wandb>=0.16.0
sentencepiece>=0.1.99
protobuf>=4.25.1
scipy>=1.11.3
einops>=0.7.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
langchain-openai>=0.0.5
langchain-core>=0.1.0
pydantic>=2.5.2
tenacity>=8.2.3
openai>=1.3.7
